third_party/blink/web_tests/svg/parser/foreign-object-case-sensitivity-expected.txt

Testing case-sensitivity for the svg:foreignObject tag in html.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Node.localName is normalized to camel-case inside <svg>, lower-case otherwise.

PASS fo1.localName is "foreignobject"
PASS fo2.localName is "foreignobject"
PASS fo3.localName is "foreignObject"
PASS fo4.localName is "foreignObject"

Selectors API queries match case-insensitively for html and svg elements in html documents.
(Should have matched case-sensitively for svg elements according to the html spec.)

PASS queryAllLower.length is 4
PASS queryAllLower[0].id is "fo1"
PASS queryAllLower[1].id is "fo2"
PASS queryAllLower[2].id is "fo3"
PASS queryAllLower[3].id is "fo4"
PASS queryAllCamel.length is 4
PASS queryAllCamel[0].id is "fo1"
PASS queryAllCamel[1].id is "fo2"
PASS queryAllCamel[2].id is "fo3"
PASS queryAllCamel[3].id is "fo4"

getElementsByTagName matches case-insensitively for html elements, case-sensitively for svg elements in html documents.

PASS byTagNameLower.length is 2
PASS byTagNameLower[0].id is "fo1"
PASS byTagNameLower[1].id is "fo2"
PASS byTagNameCamel.length is 4
PASS byTagNameCamel[0].id is "fo1"
PASS byTagNameCamel[1].id is "fo2"
PASS byTagNameCamel[2].id is "fo3"
PASS byTagNameCamel[3].id is "fo4"

Style rule matches case-insensitively for html and svg elements in html documents.
(Should have matched case-sensitively for svg elements according to the html spec).

PASS getComputedStyle(fo1).color is "rgb(0, 128, 0)"
PASS getComputedStyle(fo1).fontWeight is "700"
PASS getComputedStyle(fo2).color is "rgb(0, 128, 0)"
PASS getComputedStyle(fo2).fontWeight is "700"
PASS getComputedStyle(fo3).color is "rgb(0, 128, 0)"
PASS getComputedStyle(fo3).fontWeight is "700"
PASS getComputedStyle(fo4).color is "rgb(0, 128, 0)"
PASS getComputedStyle(fo4).fontWeight is "700"
PASS successfullyParsed is true

TEST COMPLETE

This text should be green and bold
This text should be green and bold
This text should be green, not bold
This text should be green, not bold
